REFACTORING TO SOC - runs
Docker build and run but no streamlit
