+++ modules/get_resource_manager.py
@@ -1,0 +1,6 @@
+# modules.get_resource_manager.py
+from modules.rest_modules.rest_utils.resource_manager import resource_manager
+
+
+async def get_resource_manager():
+    return resource_manager
