+++ modules/rest_modules/endpoints/health.py
@@ -3,16 +3,13 @@ from fastapi.responses import JSONResponse
 
 router = APIRouter()
 
-
 @router.get("/health")
 async def health_check():
     return JSONResponse(content={"status": "healthy"})
 
-
 @router.get("/")
 async def read_root():
     return {"message": "Welcome to the Grainger Recommendations API"}
-
 
 @router.get("/favicon.ico")
 async def favicon():
